Hyper-Dev session 1 (of 1?).
I basically rewrote the whole app.  It now requests data from the server
asynchronously and displays a progress dialog letting you know it's
working.  This prevents the HTTP request and possibly the load from the file
cache from blocking the UI thread like it would do if say you had very poor
network connection and the request was taking awhile.

You will notice the addition of GetMenuTask.java and the enclosed Result
class.  I tried to seperate the loading of the data from the presentation
of the data.  This is more efficient and easier to use.  The whole app is
much more readable now.  This is also the standar engineering choice in
larger applications.  Some even utilize their own service rather than simply
a task thread.

The main activity instantiates GetMenuTask and executes it.  It keeps a
referrence to this task in one of the member fields so that two aren't
started at once.

Once the task finishes, it calles the appropriate callback method in the UI
thread as defined by the custom listener interface I wrote.  Because of this,
the main Activity handles most of the error checking and failing gracefully.
This is better than the randomly interspersed error and bounds checking that
the previous version contained.

Other minor changes include:
	- Reworked Log entries.  Changed the priority to interactive rather
	  than debug.
	- Moved methods around to make the source more readable.
	- Visually  optimized the member fields of the main activity class
	  and removed unecessary one.
	- Renamed a few methods so they make more sense.
	- Commented almost everything.
	- Took care of like 30 TODO:'s
	- Reworked the lifecycle a bit.
	- Removed the "All" option from the preferences.
	- Reworked the old populateEntrees so it doesn't show Toast but
	  instead returns a key into the showToast method.

Filters are still untested as are the sharedPreferences.  Also, the
onSaveInstanceState and onRestoreInstanceState methods are not really tested.
They aren't used that much so it's low priority but they should be ironed out
eventually.